requirements.txt
numpy>=1.16.3
pandas>=0.24.2
scikit-learn>=0.21.1
scipy>=1.3.0
torch>=1.1.0
umap-learn>=0.3.8
universal-divergence>=0.2.0